% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{discretize}
\alias{discretize}
\title{Discretize a continuous variable}
\usage{
discretize(x, probs = c(0.25, 0.5, 0.75, 1))
}
\arguments{
\item{x}{numeric vector to convert}

\item{probs}{numeric vector of quantiles that determine factor levels}
}
\value{
factor with levels base on quantiles
}
\description{
Converts a continuous variable into a factor with levels based on quantiles of the continuous variable
}
\examples{
discretize(rnorm(20))
}
