% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tethered.R
\name{tethered_lm}
\alias{tethered_lm}
\title{Tethered hacking interval for linear model}
\usage{
tethered_lm(mdl, theta = 0.05)
}
\arguments{
\item{mdl}{model}

\item{theta}{numeric}
}
\value{
vector
}
\description{
Tethered hacking interval for linear model
}
