% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackint.R
\name{hackint_lm}
\alias{hackint_lm}
\title{Hacking intervals for linear models}
\usage{
hackint_lm(mdl, data, theta = 0.05, frac_remove_obs = 1,
  verbose = TRUE)
}
\arguments{
\item{mdl}{model}

\item{data}{dataframe}

\item{theta}{numeric}

\item{frac_remove_obs}{numeric}

\item{verbose}{boolean}
}
\value{
list
}
\description{
Hacking intervals for linear models
}
