% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tethered.R
\name{tethered_lm}
\alias{tethered_lm}
\title{Tethered hacking interval for linear model}
\usage{
tethered_lm(mdl, theta = 0.1)
}
\arguments{
\item{mdl}{\code{lm} object representing "base" model}

\item{theta}{loss tolerance for tethered hacking (default = 0.1)}
}
\value{
vector with lower bound and upper bound of hacking interval (\code{LB} and \code{UB})
as well as minimum loss estimate (\code{Estimate}).
}
\description{
Tethered hacking interval for linear model
}
